Runtime should not be shared between generator runs
