+++ machine/v1beta1/0000_10_machinehealthcheck.yaml
@@ -120,6 +120,7 @@ spec:
                       type: object
                       additionalProperties:
                         type: string
+                  x-kubernetes-map-type: atomic
                 unhealthyConditions:
                   description: UnhealthyConditions contains a list of the conditions that determine whether a node is considered unhealthy.  The conditions are combined in a logical OR, i.e. if any of the conditions is met, the node is unhealthy.
                   type: array
