+++ machine/v1beta1/0000_10_machineset.crd.yaml
@@ -102,6 +102,7 @@ spec:
                       type: object
                       additionalProperties:
                         type: string
+                  x-kubernetes-map-type: atomic
                 template:
                   description: Template is the object that describes the machine that will be created if insufficient replicas are detected.
                   type: object
@@ -159,6 +160,7 @@ spec:
                               uid:
                                 description: 'UID of the referent. More info: http://kubernetes.io/docs/user-guide/identifiers#uids'
                                 type: string
+                            x-kubernetes-map-type: atomic
                     spec:
                       description: 'Specification of the desired behavior of the machine. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status'
                       type: object
@@ -268,6 +270,7 @@ spec:
                                   uid:
                                     description: 'UID of the referent. More info: http://kubernetes.io/docs/user-guide/identifiers#uids'
                                     type: string
+                                x-kubernetes-map-type: atomic
                         providerID:
                           description: ProviderID is the identification ID of the machine provided by the provider. This field must match the provider ID as seen on the node object corresponding to this machine. This field is required by higher level consumers of cluster-api. Example use case is cluster autoscaler with cluster-api as provider. Clean-up logic in the autoscaler compares machines to nodes to find out machines at provider which could not get registered as Kubernetes nodes. With cluster-api as a generic out-of-tree provider for autoscaler, this field is required by autoscaler to be able to have a provider view of the list of machines. Another list of nodes is queried from the k8s apiserver and then a comparison is done to find out unregistered machines and are marked for delete. This field will be set by the actuators and consumed by higher level entities like autoscaler that will be interfacing with cluster-api as generic provider.
                           type: string
